Add model autodetection
Different device models (even when supported by a single integration) can have different features or even different properties/API.
Previously, some integrations have used a model argument to define the exact device type to be able to adapt to these differences,
but there has been no generalized solution for this issue.

This PR will introduce automatic model detection based on the miIO.info query.
The first invokation of any command-decorated method will do the query to find the model of the device.
The response of this query is cached, so this will only happen once.

* info() has now a new keyword-only argument 'skip_cache' which can be used to bypass the cache
* Device constructor has a new keyword-only argument to specifying the model (which skips the info query)
* This PR converts Vacuum class to use these new facilities for fanspeed controls